+++ .gitignore
@@ -30,5 +30,3 @@ tmp/
 tmp*.py
 
 .DS_Store
-
-test*.py
