+++ dolfyn/test/data/vector_data01_rotate_inst2earth.h5
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:b091d77796a0a8a87fbd4ee641811da91891c643f6dc05df17fa4d709627b02f
+size 154680
